+++ src/app/src/components/FacilityDetailSidebarExtended.jsx
@@ -11,6 +11,7 @@ import moment from 'moment';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import List from '@material-ui/core/List';
 
+import FacilityDetailSidebarClosureStatus from './FacilityDetailSidebarClosureStatus';
 import FacilityDetailsStaticMap from './FacilityDetailsStaticMap';
 import FacilityDetailSidebarHeader from './FacilityDetailSidebarHeader';
 import FacilityDetailSidebarItem from './FacilityDetailSidebarItem';
@@ -19,15 +20,19 @@ import FacilityDetailSidebarContributors from './FacilityDetailSidebarContributo
 import FacilityDetailSidebarAction from './FacilityDetailSidebarAction';
 import ReportFacilityStatus from './ReportFacilityStatus';
 import ShowOnly from './ShowOnly';
+import FeatureFlag from './FeatureFlag';
 
 import {
     fetchSingleFacility,
     resetSingleFacility,
+    fetchFacilities,
 } from '../actions/facilities';
 
 import {
     facilitySidebarActions,
     EXTENDED_FIELD_TYPES,
+    REPORT_A_FACILITY,
+    FACILITIES_REQUEST_PAGE_SIZE,
 } from '../util/constants';
 
 import {
@@ -126,6 +131,36 @@ const filterByUniqueField = (data, extendedFieldName) =>
         item => item.primary + item.secondary,
     );
 
+const formatActivityReports = data => {
+    const reports = get(data, 'properties.activity_reports', []);
+    if (!reports.length) return [null, []];
+    const formattedReports = reports.reduce((list, r) => {
+        let updatedList = [...list];
+        if (r.status === 'CONFIRMED') {
+            updatedList = [
+                ...updatedList,
+                {
+                    primary: `Verified ${r.closure_state.toLowerCase()}`,
+                    secondary: formatAttribution(r.status_change_date),
+                    key: `${r.id}-verified}`,
+                },
+            ];
+        }
+        return [
+            ...updatedList,
+            {
+                primary: `Reported ${r.closure_state.toLowerCase()}`,
+                secondary: formatAttribution(
+                    r.created_at,
+                    r.reported_by_contributor,
+                ),
+                key: r.id,
+            },
+        ];
+    }, []);
+    return [formattedReports[0], formattedReports.slice(1)];
+};
+
 const FacilityDetailSidebar = ({
     classes,
     data,
@@ -143,6 +178,8 @@ const FacilityDetailSidebar = ({
     facilityIsClaimedByCurrentUser,
     embedContributor,
     embedConfig,
+    searchForFacilities,
+    vectorTileFlagIsActive,
 }) => {
     useEffect(() => {
         fetchFacility(Number(embed), contributors);
@@ -198,6 +235,11 @@ const FacilityDetailSidebar = ({
         }
         return [defaultAddressField[0], otherAddressFields];
     }, [data]);
+
+    const [activityReport, otherActivityReports] = useMemo(
+        () => formatActivityReports(data),
+        [data],
+    );
 
     if (fetching) {
         return (
@@ -312,6 +354,14 @@ const FacilityDetailSidebar = ({
                     push={push}
                     oarId={data.properties.oar_id}
                     onClaimFacility={claimFacility}
+                    onBack={() => {
+                        clearFacility();
+                        searchForFacilities(vectorTileFlagIsActive);
+                    }}
+                />
+                <FacilityDetailSidebarClosureStatus
+                    data={data}
+                    clearFacility={clearFacility}
                 />
                 <FacilityDetailSidebarItem
                     label="OAR ID"
@@ -346,6 +396,16 @@ const FacilityDetailSidebar = ({
                     />
                     {EXTENDED_FIELD_TYPES.map(renderExtendedField)}
                 </ShowOnly>
+                <FeatureFlag flag={REPORT_A_FACILITY}>
+                    <ShowOnly when={!!activityReport}>
+                        <FacilityDetailSidebarItem
+                            label="Status"
+                            {...activityReport}
+                            additionalContent={otherActivityReports}
+                            embed={embed}
+                        />
+                    </ShowOnly>
+                </FeatureFlag>
                 <ShowOnly when={embed}>{renderEmbedFields()}</ShowOnly>
                 <div className={classes.actions}>
                     <ShowOnly when={!embed}>
@@ -405,6 +465,7 @@ function mapStateToProps(
         auth: { user },
         embeddedMap: { embed, config },
         filters: { contributors },
+        featureFlags,
     },
     {
         match: {
@@ -428,6 +489,12 @@ function mapStateToProps(
         includes(currentUserPendingClaimedFacilities, oarID) &&
         !facilityIsClaimedByCurrentUser;
 
+    const vectorTileFlagIsActive = get(
+        featureFlags,
+        'flags.vector_tile',
+        false,
+    );
+
     return {
         data,
         fetching,
@@ -438,6 +505,7 @@ function mapStateToProps(
         contributors,
         userHasPendingFacilityClaim,
         facilityIsClaimedByCurrentUser,
+        vectorTileFlagIsActive,
     };
 }
 
@@ -453,6 +521,14 @@ function mapDispatchToProps(
         fetchFacility: (embed, contributorId) =>
             dispatch(fetchSingleFacility(oarID, embed, contributorId)),
         clearFacility: () => dispatch(resetSingleFacility()),
+        searchForFacilities: vectorTilesAreActive =>
+            dispatch(
+                fetchFacilities({
+                    pageSize: vectorTilesAreActive
+                        ? FACILITIES_REQUEST_PAGE_SIZE
+                        : 50,
+                }),
+            ),
     };
 }